NEXUS-22136 REST API for Helm repositories
